More testAdapter
